Remote access M2: shared notification core + daemon-side engine
Move the pure notification heart out of the TUI into repomon-core::notify
(NotifKind, SessKey, per-session status keying, transition diffing,
compose/compose_burst, new slot_by_key helper) — tests move with it; the
TUI re-imports and keeps only delivery (popups, banners, feed). NotifKind
gains snake_case serde for wire payloads.

New daemon task notify_watch: every 4s (self-gated on [remote] enabled +
notify_enabled, so it costs nothing for TUI-only use and reacts to live
config flips) it diffs the lane list with the shared detection, debounces
per (lane, session, kind), and broadcasts event.notification
{lane_id, session_id, kind, title, body, prompt} to subscribed clients —
the feed a companion app renders and the hook APNs sends from (M3).
lane.list's composite moved to a shared lanes_with_agents helper.
