+++ crates/repomon-tui/src/app.rs
@@ -16,6 +16,9 @@ use repomon_core::model::{
     AgentChoice, AgentSession, AgentStatus, BrowseEntry, BrowseResult, Commit, Lane, LaneId, Repo,
     TimelineData, WorkSession,
 };
+use repomon_core::notify::{
+    diff_session_transitions, session_by_key, session_statuses, slot_by_key, SessKey,
+};
 use repomon_core::protocol::Notification;
 use serde_json::json;
 use tokio::sync::{broadcast, mpsc};
@@ -703,12 +706,7 @@ impl App {
         let Some((title, body, session_id)) = self.lanes.iter().find(|l| l.id == id).map(|l| {
             let sess = session_by_key(l, key);
             // Which of the lane's side-by-side agents this is (1-based in the body when >1).
-            let slot = sess.and_then(|hit| {
-                let real: Vec<&AgentSession> =
-                    l.agent_sessions.iter().filter(|s| !s.inferred).collect();
-                let i = real.iter().position(|s| std::ptr::eq(*s, hit))?;
-                Some((i, real.len()))
-            });
+            let slot = slot_by_key(l, key);
             let (t, b) = notify::compose(kind, l, sess, slot, self.settings.notify_show_why);
             (t, b, sess.and_then(|s| s.session_id.clone()))
         }) else {
@@ -2906,122 +2904,6 @@ fn fuzzy_score(haystack: &str, needle: &str) -> Option<u32> {
     Some(score)
 }
 
-/// Identifies one real agent session within a lane across refreshes.
-///
-/// Transcript-backed sessions key on the Claude session id (the transcript filename stem),
-/// which is stable across polls. `claude --resume` may continue the same logical work in a new
-/// transcript; that reads as one session vanishing and another appearing — acceptable noise. A
-/// lane has at most one real session *without* a transcript id per snapshot (the managed
-/// no-transcript placeholder or the generic process monitor — mutually exclusive branches in
-/// the daemon's `overlay_agents`), so a single `Fallback` sentinel covers it.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-enum SessKey {
-    Transcript(String),
-    Fallback,
-}
-
-/// Key/status pairs for one lane's *real* agent sessions. Inferred "file activity" placeholders
-/// are dropped so they never drive named alerts. On a (theoretically impossible) duplicate key,
-/// the higher-priority status wins — the same order the old per-lane rollup used.
-fn session_statuses(
-    lane_id: LaneId,
-    sessions: &[AgentSession],
-) -> Vec<((LaneId, SessKey), AgentStatus)> {
-    let mut out: Vec<((LaneId, SessKey), AgentStatus)> = Vec::new();
-    for s in sessions.iter().filter(|s| !s.inferred) {
-        let key = (
-            lane_id,
-            s.session_id
-                .clone()
-                .map(SessKey::Transcript)
-                .unwrap_or(SessKey::Fallback),
-        );
-        match out.iter_mut().find(|(k, _)| *k == key) {
-            Some((_, st)) if status_priority(s.status) < status_priority(*st) => *st = s.status,
-            Some(_) => {}
-            None => out.push((key, s.status)),
-        }
-    }
-    out
-}
-
-/// Notification priority of a status (lower = more urgent).
-fn status_priority(s: AgentStatus) -> usize {
-    use AgentStatus::*;
-    [RateLimited, Waiting, Running, Idle, Ended]
-        .iter()
-        .position(|&x| x == s)
-        .unwrap_or(usize::MAX)
-}
-
-/// Diff the previous and current per-session status maps into the notifications to fire.
-///
-/// Sessions present in `now` are edge-detected against their previous status. Sessions that
-/// vanished fire as a transition to `None` (→ Idle if they were active), except when their
-/// whole lane is gone (deleting a lane isn't an agent going quiet) or when a lane's `Fallback`
-/// key was handed off to a transcript-backed key (`lanes_with_managed`): the managed
-/// no-transcript placeholder disappears the moment the agent's transcript becomes parseable,
-/// and firing Idle there would alert on every spawn.
-fn diff_session_transitions(
-    prev: &HashMap<(LaneId, SessKey), AgentStatus>,
-    now: &HashMap<(LaneId, SessKey), AgentStatus>,
-    live_lanes: &HashSet<LaneId>,
-    lanes_with_managed: &HashSet<LaneId>,
-) -> Vec<((LaneId, SessKey), NotifKind)> {
-    let mut out = Vec::new();
-    for (key, &status) in now {
-        let was = prev.get(key).copied();
-        if was == Some(status) {
-            continue;
-        }
-        if let Some(kind) = transition_kind(was, Some(status)) {
-            out.push((key.clone(), kind));
-        }
-    }
-    for (key, &was) in prev {
-        if now.contains_key(key) || !live_lanes.contains(&key.0) {
-            continue;
-        }
-        if key.1 == SessKey::Fallback && lanes_with_managed.contains(&key.0) {
-            continue;
-        }
-        if let Some(kind) = transition_kind(Some(was), None) {
-            out.push((key.clone(), kind));
-        }
-    }
-    out
-}
-
-/// Resolve a session key back to the lane's session, for composing the notification text.
-/// `None` when the session vanished (i.e. its disappearance was the trigger).
-fn session_by_key<'a>(lane: &'a Lane, key: &SessKey) -> Option<&'a AgentSession> {
-    lane.agent_sessions
-        .iter()
-        .filter(|s| !s.inferred)
-        .find(|s| match key {
-            SessKey::Transcript(id) => s.session_id.as_deref() == Some(id.as_str()),
-            SessKey::Fallback => s.session_id.is_none(),
-        })
-}
-
-/// Map a session's status transition to the notification it should fire, if any. `None` means
-/// the session was absent from that snapshot. Priority resolves cases like
-/// `Running → RateLimited` to the limit.
-fn transition_kind(prev: Option<AgentStatus>, now: Option<AgentStatus>) -> Option<NotifKind> {
-    use AgentStatus::*;
-    match (prev, now) {
-        // Hit a usage limit.
-        (p, Some(RateLimited)) if p != Some(RateLimited) => Some(NotifKind::RateLimited),
-        // Auto-resumed after a limit.
-        (Some(RateLimited), Some(Running)) => Some(NotifKind::Resumed),
-        // Finished its turn / needs you.
-        (p, Some(Waiting)) if p != Some(Waiting) => Some(NotifKind::NeedsYou),
-        // Was active, now quiet (idle / ended / the session went away).
-        (Some(Running) | Some(Waiting), Some(Idle) | Some(Ended) | None) => Some(NotifKind::Idle),
-        _ => None,
-    }
-}
-
 /// Cycle `current` to the next/previous option (wrapping). Returns `current` if `options` is empty.
 fn cycle(options: &[&str], current: &str, forward: bool) -> String {
     if options.is_empty() {
@@ -3387,44 +3269,6 @@ async fn next_note(rx: &mut broadcast::Receiver<Notification>) -> Option<Notific
 mod tests {
     use super::*;
 
-    #[test]
-    fn notification_transitions() {
-        use AgentStatus::*;
-        // The headline alerts.
-        assert_eq!(
-            transition_kind(Some(Running), Some(Waiting)),
-            Some(NotifKind::NeedsYou)
-        );
-        assert_eq!(
-            transition_kind(None, Some(Waiting)),
-            Some(NotifKind::NeedsYou)
-        );
-        assert_eq!(
-            transition_kind(Some(Running), Some(RateLimited)),
-            Some(NotifKind::RateLimited)
-        );
-        assert_eq!(
-            transition_kind(Some(RateLimited), Some(Running)),
-            Some(NotifKind::Resumed)
-        );
-        // Gave up on the limit and now needs you.
-        assert_eq!(
-            transition_kind(Some(RateLimited), Some(Waiting)),
-            Some(NotifKind::NeedsYou)
-        );
-        // Went quiet (idle / ended / the agent went away).
-        assert_eq!(
-            transition_kind(Some(Running), Some(Idle)),
-            Some(NotifKind::Idle)
-        );
-        assert_eq!(transition_kind(Some(Waiting), None), Some(NotifKind::Idle));
-        // Non-events: you replied, work simply started, or nothing changed.
-        assert_eq!(transition_kind(Some(Waiting), Some(Running)), None);
-        assert_eq!(transition_kind(None, Some(Running)), None);
-        assert_eq!(transition_kind(Some(Idle), Some(Running)), None);
-        assert_eq!(transition_kind(Some(Waiting), Some(Waiting)), None);
-    }
-
     /// A minimal real-or-inferred session, mirroring the daemon's `overlay_agents` literals.
     fn sess(session_id: Option<&str>, status: AgentStatus, inferred: bool) -> AgentSession {
         AgentSession {
@@ -3446,111 +3290,6 @@ mod tests {
             tmux_window: None,
             last_message: None,
         }
-    }
-
-    #[test]
-    fn session_statuses_keys_and_filters() {
-        use AgentStatus::*;
-        let sessions = vec![
-            sess(Some("a"), Waiting, false),
-            sess(Some("b"), RateLimited, false),
-            sess(None, Running, true), // inferred file-activity placeholder — excluded
-            sess(None, Running, false),
-        ];
-        let got = session_statuses(7, &sessions);
-        assert_eq!(got.len(), 3);
-        assert!(got.contains(&((7, SessKey::Transcript("a".into())), Waiting)));
-        assert!(got.contains(&((7, SessKey::Transcript("b".into())), RateLimited)));
-        assert!(got.contains(&((7, SessKey::Fallback), Running)));
-
-        // Defensive: a duplicate key keeps the higher-priority status.
-        let dup = vec![sess(None, Idle, false), sess(None, Waiting, false)];
-        assert_eq!(
-            session_statuses(7, &dup),
-            vec![((7, SessKey::Fallback), Waiting)]
-        );
-    }
-
-    #[test]
-    fn two_sessions_fire_independent_streams() {
-        use AgentStatus::*;
-        let k = |id: &str| (1, SessKey::Transcript(id.into()));
-        let live: HashSet<LaneId> = [1].into();
-        let managed = HashSet::new();
-
-        // One agent finishes its turn while its lane-mate is still rate-limited. The old
-        // per-lane rollup saw "RateLimited" before and after and fired nothing — the masking
-        // this change exists to fix.
-        let prev: HashMap<_, _> = [(k("a"), Running), (k("b"), RateLimited)].into();
-        let now: HashMap<_, _> = [(k("a"), Waiting), (k("b"), RateLimited)].into();
-        assert_eq!(
-            diff_session_transitions(&prev, &now, &live, &managed),
-            vec![(k("a"), NotifKind::NeedsYou)]
-        );
-
-        // And the rate-limited lane-mate resumes independently.
-        let now2: HashMap<_, _> = [(k("a"), Waiting), (k("b"), Running)].into();
-        assert_eq!(
-            diff_session_transitions(&now, &now2, &live, &managed),
-            vec![(k("b"), NotifKind::Resumed)]
-        );
-    }
-
-    #[test]
-    fn disappearance_fires_idle_only_when_lane_lives() {
-        use AgentStatus::*;
-        let k = (1, SessKey::Transcript("a".into()));
-        let prev: HashMap<_, _> = [(k.clone(), Waiting)].into();
-        let now = HashMap::new();
-        let managed = HashSet::new();
-
-        let live: HashSet<LaneId> = [1].into();
-        assert_eq!(
-            diff_session_transitions(&prev, &now, &live, &managed),
-            vec![(k, NotifKind::Idle)]
-        );
-        // The whole lane went away (deleted): not an agent going quiet.
-        assert!(diff_session_transitions(&prev, &now, &HashSet::new(), &managed).is_empty());
-    }
-
-    #[test]
-    fn fallback_handoff_does_not_fire_idle() {
-        use AgentStatus::*;
-        let live: HashSet<LaneId> = [1].into();
-        let prev: HashMap<_, _> = [((1, SessKey::Fallback), Running)].into();
-        let now: HashMap<_, _> = [((1, SessKey::Transcript("a".into())), Running)].into();
-
-        // The managed spawn's transcript became parseable: Fallback hands off to Transcript
-        // within one refresh. The agent didn't stop, so nothing fires.
-        let managed: HashSet<LaneId> = [1].into();
-        assert!(diff_session_transitions(&prev, &now, &live, &managed).is_empty());
-
-        // But with no managed session left in the lane, a vanished fallback is a real stop.
-        let gone = HashMap::new();
-        assert_eq!(
-            diff_session_transitions(&prev, &gone, &live, &HashSet::new()),
-            vec![((1, SessKey::Fallback), NotifKind::Idle)]
-        );
-    }
-
-    #[test]
-    fn new_session_already_waiting_fires_needs_you() {
-        use AgentStatus::*;
-        let live: HashSet<LaneId> = [1].into();
-        let managed = HashSet::new();
-        let prev = HashMap::new();
-        let k = (1, SessKey::Transcript("a".into()));
-
-        // A second agent appearing mid-run already waiting (e.g. a parallel session that
-        // finished between refreshes) is exactly the alert the user wants.
-        let waiting: HashMap<_, _> = [(k.clone(), Waiting)].into();
-        assert_eq!(
-            diff_session_transitions(&prev, &waiting, &live, &managed),
-            vec![(k.clone(), NotifKind::NeedsYou)]
-        );
-        // Appearing already-running is just work starting; stay quiet.
-        let running: HashMap<_, _> = [(k, Running)].into();
-        assert!(diff_session_transitions(&prev, &running, &live, &managed).is_empty());
     }
 
     #[test]
